refactor: split engine.rs god module into partial impl files
Split the 4841-line src/engine.rs into 13 focused files using Rust's
multi-file impl pattern. Each file contains a single concern's impl
block for MemoryEngine. No API changes, no behavior changes — purely
structural.

Module layout:
- engine/mod.rs — struct, constructors, internal helpers, config, scope
- engine/ingest.rs — ingest, add_fact, add_facts_batch
- engine/query.rs — query, execute_query, search/store paths
- engine/bootstrap.rs — bootstrap_session, bootstrap_directory
- engine/consolidation.rs — consolidate
- engine/forgetting.rs — forget, pin_fact, unpin_fact
- engine/conflict.rs — resolve_conflict
- engine/graph.rs — link_session_facts, graph queries
- engine/inspect.rs — statistics, replay_events, explain_fact, dump_state
- engine/restore.rs — restore_json, restore_json_memory, restore_sqlite
- engine/scheduling.rs — list_due, next_due_time
- engine/resume.rs — resume_context
- engine/tests.rs — all 420 tests (unchanged)

Fixes #108
